Sort files and change the name of the plugin
