+++ lib/vagrant-gpii-ci.rb
@@ -1,9 +1,8 @@
 require "pathname"
 
-require "vagrant-gpii-ci/plugin"
-
 module VagrantPlugins
-  module Cienv
+  module GPIICi
+
     lib_path = Pathname.new(File.expand_path("../vagrant-gpii-ci", __FILE__))
     autoload :Action, lib_path.join("action")
     autoload :Errors, lib_path.join("errors")
@@ -16,3 +15,17 @@ module VagrantPlugins
     end
   end
 end
+
+begin
+  require "vagrant"
+rescue LoadError
+  raise "The Vagrant GPII CI plugin must be run within Vagrant."
+end
+
+# This is a sanity check to make sure no one is attempting to install
+# this into an early Vagrant version.
+if Vagrant::VERSION < "1.8.0"
+  raise "The Vagrant GPII CI plugin is only compatible with Vagrant 1.8+"
+end
+
+require "vagrant-gpii-ci/plugin"
