+++ lib/vagrant-gpii-ci/action/init_environment.rb
@@ -1,7 +1,7 @@
 require "log4r"
 
 module VagrantPlugins
-  module Cienv
+  module GPIICi
     module Action
 
       # This action sets up the QI QI environment
@@ -9,11 +9,10 @@ module VagrantPlugins
 
         def initialize(app, env)
           @app = app
-          @logger = Log4r::Logger.new("vagrant_qi::action::init_environment")
+          @logger = Log4r::Logger.new("VagrantPlugins::GPIICi::action::init_environment")
         end
 
         def call(env)
-          puts "i am inside class InitEnvironment"
           @logger.info('i am inside class InitEnvironment - call()')
           @app.call(env)
         end
