+++ lib/vagrant-gpii-ci/version.rb
@@ -1,5 +1,5 @@
 module VagrantPlugins
-  module Cienv
+  module GPIICi
     VERSION = "0.0.1"
   end
 end
