+++ spec/spec_helper.rb
@@ -1,2 +1,5 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
-require 'vagrant/cienv'
+require 'vagrant-gpii-ci'
+
+RSpec.configure do |spec|
+end
