--- spec/vagrant/cienv_spec.rb
@@ -1,11 +1,0 @@
-require 'spec_helper'
-
-describe Vagrant::Cienv do
-  it 'has a version number' do
-    expect(Vagrant::Cienv::VERSION).not_to be nil
-  end
-
-  it 'does something useful' do
-    expect(false).to eq(true)
-  end
-end
